fix(ui): compute gas in the dApp for batch supplier transactions
Soothe's gas:'auto'/gasAdjustment did not apply on the user's build, so a
batch stake (10 suppliers = 20 msgs) hit on-chain code 11 (out of gas) when
gas_limit landed at ~gasUsed with no margin. Move estimation into the dApp:
a shared estimateGas() (registry + sequence + simulate, gasLimit = ceil(gasUsed*2))
on the base WalletConnection; Pocket/Soothe now passes an explicit numeric gas
with gasAdjustment 1 (bypassing Soothe auto), and Keplr reuses the same estimator.
